Add validation step for possible duplicates
If file somehow changes between runs (for whatever
reason, including file corruption), the hash will change.
But the filename will not.
The record step will then add another entry to the record,
resulting in two hashes pointing to one filename.
This isn't breaking error, but still incosistent.
Notify the user of the invalid entries and ask for purge
or fix.